+++ core/src/com/unibo/util/LineOfSight.java
@@ -9,30 +9,33 @@ import com.unibo.model.Character;
 import com.unibo.view.HeroView;
 import com.unibo.view.LevelView;
 import com.unibo.view.MobView;
-
+/**
+ *
+ */
 public final class LineOfSight {
-	
-	public static boolean isHeroSeen(MobView mob, LevelView level, DescentMap map) {
-		final HeroView hero = level.getHeroView();
-		Vector2 heroVector = characterVector(mob.getCharacter());
-		Vector2 mobVector = characterVector(hero.getCharacter());
-		
-		for (final PolygonMapObject polyMapObj : map.getCollisionLayer().getObjects().getByType(PolygonMapObject.class)) {
-			Polygon polyObj = new Polygon(polyMapObj.getPolygon().getTransformedVertices());
-			var verts = polyObj.getTransformedVertices();
-	            for (int i = 0; i < verts.length; i++) {
-	                verts[i] *= map.getUnitScale();
-	            }
-			if (Intersector.intersectSegmentPolygon(mobVector, heroVector, polyObj)) {   
-				return false;
-			}
-		}
-		return true;		
-	}
-	
-	public static Vector2 characterVector(Character character) {
-		final int x = character.getPos().getxCoord();
-		final int y = character.getPos().getyCoord();
-		return new Vector2(x,y);
-	}
+
+    public static boolean isHeroSeen(MobView mob, LevelView level, DescentMap map) {
+        final HeroView hero = level.getHeroView();
+        Vector2 heroVector = characterVector(mob.getCharacter());
+        Vector2 mobVector = characterVector(hero.getCharacter());
+
+        for (final PolygonMapObject polyMapObj : map.getCollisionLayer().getObjects()
+                .getByType(PolygonMapObject.class)) {
+            Polygon polyObj = new Polygon(polyMapObj.getPolygon().getTransformedVertices());
+            var verts = polyObj.getTransformedVertices();
+            for (int i = 0; i < verts.length; i++) {
+                verts[i] *= map.getUnitScale();
+            }
+            if (Intersector.intersectSegmentPolygon(mobVector, heroVector, polyObj)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    public static Vector2 characterVector(Character character) {
+        final float x = character.getPos().getxCoord();
+        final float y = character.getPos().getyCoord();
+        return new Vector2(x, y);
+    }
 }
